# openapi.yaml
openapi: 3.1.0
info:
  description: "Flow definition management and validation executions. Every request\
    \ except the health endpoints needs an `X-API-Key` header; the key fixes the tenant\
    \ and its scopes. Operational probes are also served at `/actuator/health/liveness`\
    \ and `/actuator/health/readiness` (public, Spring Actuator format)."
  title: Validation Flow Engine API
  version: v1
servers:
- description: This server
  url: /
...
paths:
  /api/v1/executions:
    get:
      description: "Tenant's executions, paginated (`size` 1..100)."
      operationId: list_1
      parameters:
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
      security:
      - apiKey:
        - validation:read
      summary: List executions
      tags:
      - Execution
    post:
      description: Runs the ACTIVE flow synchronously. With `Idempotency-Key` (scoped
        to the tenant) a replay with the same payload returns the original execution;
        a different payload returns 409.
      operationId: start
      parameters:
      - description: Optional idempotency key (max 255 chars)
        example: order-42
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExecutionResponse"
          description: Created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
      security:
      - apiKey:
        - validation:execute
      summary: Start an execution
      tags:
      - Execution
  /api/v1/executions/{id}:
    get:
      description: Another tenant's id returns 404.
      operationId: get_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
      security:
      - apiKey:
        - validation:read
      summary: Get an execution
      tags:
      - Execution
  /api/v1/executions/{id}/nodes:
    get:
      description: "One entry per node attempt, with masked snapshots."
      operationId: nodes
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
      security:
      - apiKey:
        - validation:read
      summary: List node attempts
      tags:
      - Execution
...
components:
  responses:
    BadRequest:
      content:
        application/json:
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                details:
                - "displayName: must not be blank"
                retryable: false
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: "Invalid request (VALIDATION_ERROR, MALFORMED_REQUEST, BAD_REQUEST,\
        \ INVALID_INPUT)."
    Conflict:
      content:
        application/json:
          examples:
            default:
              value:
                code: CONFLICT
                details:
                - flow is not DRAFT
                retryable: false
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: "Conflict (CONFLICT): duplicate resource, wrong lifecycle state,\
        \ or Idempotency-Key reused with a different payload."
    Forbidden:
      content:
        application/json:
          examples:
            default:
              value:
                code: FORBIDDEN
                details:
                - missing scope
                retryable: false
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: The key lacks the required scope.
    NotFound:
      content:
        application/json:
          examples:
            default:
              value:
                code: FLOW_NOT_FOUND
                details:
                - flow not found
                retryable: false
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: "Resource not found, or it belongs to another tenant (FLOW_NOT_FOUND,\
        \ EXECUTION_NOT_FOUND)."
    Unauthorized:
      content:
        application/json:
          examples:
            default:
              value:
                code: UNAUTHORIZED
                details:
                - authentication required
                retryable: false
          schema:
            $ref: "#/components/schemas/ErrorResponse"
      description: Missing or invalid `X-API-Key`.
...
  schemas:
...
    ErrorResponse:
      description: Error body used by every non-2xx response.
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: FLOW_NOT_FOUND
        details:
          type: array
          description: Human-readable details
          example:
          - flow not found
          items:
            type: string
            description: Human-readable details
            example: "[\"flow not found\"]"
        retryable:
          type: boolean
          description: True when retrying the same request may succeed
    ExecutionResponse:
      type: object
      properties:
        completedAt:
          type: string
          format: date-time
        error:
          type: object
          additionalProperties: {}
          description: Error info when FAILED (dynamic)
          example:
            code: VALIDATOR_FAILED
        executionId:
          type: string
          format: uuid
        flowKey:
          type: string
          example: onboarding
        flowVersion:
          type: integer
          format: int32
          example: 1
        result:
          type: object
          additionalProperties: {}
          description: Final context (dynamic); credential-named keys are masked
          example:
            nodes:
              start: {}
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          example: COMPLETED
...
    StartRequest:
      type: object
      properties:
        context:
          type: string
          example: ONBOARDING
          minLength: 1
        correlationId:
          type: string
          example: req-123
        flowKey:
          type: string
          description: "Optional: pick the ACTIVE flow by key instead of userType+context"
          example: onboarding
        inputData:
          type: object
          additionalProperties: {}
          description: "Dynamic input, validated against the flow's inputContract"
          example:
            cpf: "12345678900"
        userType:
          type: string
          example: PF
          minLength: 1
      required:
      - context
      - inputData
      - userType
...
  securitySchemes:
    apiKey:
      description: Static API key. Operations list the scope they require (403 when
        missing).
      in: header
      name: X-API-Key
      type: apiKey